盲盒上线后通用常量的重构

+++ nft-turbo-common/nft-turbo-api/src/main/java/cn/hollis/nft/turbo/api/collection/response/CollectionInventoryModifyResponse.java
@@ -20,5 +20,5 @@ public class CollectionInventoryModifyResponse extends CollectionModifyResponse 
     /**
      * 修改的数量
      */
-    private Long quantityModified;
+    private Integer quantityModified;
 }
